:art: Rename data model names
