+++ applications/backend/src/db/repositories/sqlalchemy/base_repository.py
@@ -3,7 +3,7 @@ from typing import Generic, Type, TypeVar
 from pydantic import BaseModel
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from src.db.data_models import (
+from db.sqlalchemy_data_models import (
     SQLAlchemyDeck,
     SQLAlchemyGenre,
     SQLAlchemyItem,
