+++ applications/backend/src/db/repositories/sqlalchemy/test_item_repository.py
@@ -1,7 +1,7 @@
 from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from src.db.data_models import SQLAlchemyTestItem
+from db.sqlalchemy_data_models import SQLAlchemyTestItem
 from src.domain.models import TestItem
 
 from .base_repository import BaseRepository
